requirements.txt
alembic
aiogram
pydantic
pydantic[email]
pyopenssl
httpx
python-dotenv
psycopg2-binary
flask-wtf
flask
flask-login
sqlalchemy
sqlalchemy-utils
sqlalchemy-mixins
werkzeug
flask-mail
flask-cors